services/query-engine/pyproject.toml
[project]
name = "agent-query-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "Laurence Sonnenberg",email = "laurence.sonnenberg@acses.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.44.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "llama-index-core (>=0.14.18,<0.15.0)",
    "llama-index-llms-openai (>=0.7.3,<0.8.0)",
    "llama-index-embeddings-openai (>=0.6.0,<0.7.0)",
    "llama-index-graph-stores-neo4j (>=0.7.0,<0.8.0)",
    "llama-index-readers-file (>=0.6.0,<0.7.0)",
    "llama-index-readers-gcs (>=0.6.1,<0.7.0)",
    "celery[redis] (>=5.6.2,<6.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "llama-index-vector-stores-postgres (>=0.8.1,<0.9.0)",
    "llama-index-storage-docstore-postgres (>=0.5.0,<0.6.0)",
    "llama-index-storage-index-store-postgres (>=0.6.0,<0.7.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "docx2txt (>=0.9,<0.10)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.15.7,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "testcontainers[neo4j,postgres,redis] (>=4.0.0,<5.0.0)",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: requires Docker (Neo4j, PostgreSQL, Redis)",
]
filterwarnings = [
    "ignore::DeprecationWarning:testcontainers",
    "ignore:Using a driver after it has been closed:DeprecationWarning",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["B008"]  # Depends() and Query() in FastAPI signatures are intentional

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "llama_index.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "celery.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.worker.tasks"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

